Further modifications of late 2007

+++ plugins/content/class.cArticleLayout.php
@@ -1,10 +1,20 @@
-<?php
+<?php if(!defined('BORDERLESS')) { header('Location: / ',true,403); exit(); }
+/* Borderless CMS - the easiest and most flexible way to a valid website
+ *   (c) 2004-2007 Alexander Heusingfeld <aheusingfeld@borderlesscms.de>
+ *   Distributed under the terms and conditions of the GPL as stated in /license.txt
+ * EXCLUSION:
+ *   The files in the folder /pear/* are part of the PHP PEAR Project and are therefore
+ *   distributed under the terms and conditions of the PHP License as stated in /pear/LICENSE
+ */
+
 /**
+ * Handles and creates article write and edit form
  *
- *
- * @author ahe <aheusingfeld@borderless-cms.de>
- * @package layout
- * @version $Id$
+ * @since 0.9
+ * @author ahe <aheusingfeld@borderlessscms.de>
+ * @class cArticleLayout
+ * @ingroup content
+ * @package content
  */
  class cArticleLayout extends Layout_DAL {
 
@@ -25,7 +35,7 @@
 	$formArray = $form->toArray();
 
 	$p_sLayout = '<form  id="form_contenttext" name="form_contenttext" '
-	  .'action="'.BcmsFactory::getInstanceOf('Parser')->getServerParameter('SCRIPT_URI').'" method="post" enctype="'
+	  .'action="'.BcmsSystem::getParser()->getServerParameter('SCRIPT_URI').'" method="post" enctype="'
 	  .BcmsConfig::getInstance()->default_form_enctype
 	  ."\">\n".$p_sLayout;
 	for ($index = 0; $index < count($p_aFields); $index++) {
@@ -38,7 +48,7 @@
 		// the image shall be displayed...
 		$imageID =
 		  is_numeric($_SESSION['current_article_data']['content'][$tech_title]) ? $_SESSION['current_article_data']['content'][$tech_title] : $p_aFields[$index]['preset_value'];
-		$bcmsObject = PluginManager::getPlgInstance('ObjectManager');
+		$bcmsObject = PluginManager::getPlgInstance('FileManager');
 		$imgTag = $bcmsObject->getObjectsSmallImage($imageID);
 
 		// ... and a hidden form field with the index number
@@ -64,12 +74,17 @@
 		$formElement,$p_sLayout);
 	}
 	// ...and at last the submit button
-	$element = $form->addElement('submit','article_update_preview');
-	$element->setValue(Factory::getObject('Dictionary')->getTrans('Preview'));
+	$element = $form->addElement('submit','article_update_preview', BcmsSystem::getDictionaryManager()->getTrans('Preview'));
 	$p_sLayout .= $element->toHtml();
-	$element = $form->addElement('submit',$submit_button_name);
-	$element->setValue(Factory::getObject('Dictionary')->getTrans('continue'));
+
+	$form->addElement('submit', 'abort_action', BcmsSystem::getDictionaryManager()->getTrans('cancel'));
+	$element = $form->getElement('abort_action');
+	$element->setLabel('&nbsp;');
+	$p_sLayout .= $element->toHtml();
+	
+	$element = $form->addElement('submit',$submit_button_name, BcmsSystem::getDictionaryManager()->getTrans('continue'));
 	$p_sLayout .= $element->toHtml()."\n".'</form>';
+	
 	return $p_sLayout;
    }
 
@@ -79,7 +94,7 @@
 		$tech_title, 'title='.$tech_title);
 
 		// check whether the form field is given in the article_data
-		if(isset($_SESSION['current_article_data']['content'][$tech_title]))
+		if(!empty($_SESSION['current_article_data']['content'][$tech_title]) ) 
 		{
 			// ... if so set this to the current formfields value
 			$element->setValue(stripslashes(
@@ -104,25 +119,31 @@
    }
 
 	private function phase2(&$p_refArticle, $submit_button_name) {
+		if(!array_key_exists('current_article_data',$_SESSION)) {
+			return BcmsSystem::raiseError('Formulardaten von Phase1 verloren! Bitte vermeiden '.
+				' Sie während der Erstellung von Beiträgen einen Wechsel zu anderen '. // @todo use dictionary!
+				' Seiten dieses Angebots.','phase2()',__FILE__, __LINE__);
+		}
 		// add the form elements for the current layout
 		$layoutData =
 			$this->getLayoutDataFromFS($_SESSION['current_article_data']['layout_id']);
 		$layout = $layoutData[0]; // Achtung: Das CSS wird hier noch nicht geholt!
-		$lfields =
-			$this->getLayoutFields($_SESSION['current_article_data']['layout_id']);
-		// URGENT check where the following comes from
+		$lfields = $this->getLayoutFields($_SESSION['current_article_data']['layout_id']);
+			
+		// \bug URGENT check where the following comes from
 		if(empty($_SESSION['current_article_data']['layout_id']))
 			throw new Exception('$_SESSION does not contain layout_id!!!');
 
 		$lfields = $this->setChosenImageToArray($lfields);
 		$lfields = $this->presetValuesFromFirstPage($lfields);
-		echo Factory::getObject('GuiUtility')->createHeading(3,'Phase2: Artikel schreiben'); // TODO use dictionary
-		$guide_text = Factory::getObject('Dictionary')
-			->getTrans('write_article_info2');
+		echo BcmsFactory::getInstanceOf('GuiUtility')->createHeading(3,'Phase2: Artikel schreiben'); // @todo use dictionary
+
+		$guide_text = BcmsSystem::getDictionaryManager()->getTrans('write_article_info2');
 		echo '<div id="guide_text">'.stripcslashes($guide_text).'</div>';
-		$editor_hints = Factory::getObject('Dictionary')
-			->getTrans('editor_syntax_hints');
+		
+		$editor_hints = BcmsSystem::getDictionaryManager()->getTrans('editor_syntax_hints');
 		echo '<div id="formatting_info">'.stripcslashes($editor_hints).'</div>';
+
 		// print out the article preview
 		echo '<div id="article_preview_surrounder" class="floatclear">',"\n";
 		echo $this->parseFieldsIntoLayoutForm($layout, $lfields, $submit_button_name);
@@ -152,8 +173,8 @@
 		if($lfields[0]['tech_title']=='heading'
 			&& !empty($_POST['heading']))
 		{
-			$lfields[0]['preset_value'] =
-				BcmsFactory::getInstanceOf('Parser')->getPostParameter('heading');
+			$lfields[0]['preset_value'] = stripslashes(
+				BcmsSystem::getParser()->getPostParameter('heading'));
 		}
 		$notFound = true;
 		$i=0;
@@ -162,7 +183,7 @@
 				&& !empty($_POST['description']))
 			{
 				$lfields[$i]['preset_value'] =
-					BcmsFactory::getInstanceOf('Parser')->getPostParameter('description');
+					BcmsSystem::getParser()->getPostParameter('description');
 				$notFound=false;
 			}
 			$i++;
@@ -171,29 +192,27 @@
 	}
 
 	private function phase1(&$p_refArticle, $submit_button_name) {
-		echo PluginManager::getPlgInstance('CategoryManager')->getLogic()->createHeading(3,'Phase1: Artikel-Metainformationen pflegen');
-		$guide_text = Factory::getObject('Dictionary')
+		echo BcmsSystem::getCategoryManager()->getLogic()->createHeading(3,'Phase1: Artikel-Metainformationen pflegen');
+		$guide_text = BcmsSystem::getDictionaryManager()
 			->getTrans('write_article_info1');
 		echo '<div id="guide_text">'.stripcslashes($guide_text).'</div>';
 		$form =& $p_refArticle->getForm('article_form',$submit_button_name
-		  ,Factory::getObject('Dictionary')->getTrans('continue'));
+		  ,BcmsSystem::getDictionaryManager()->getTrans('continue'));
 		$form->validate(); // validate entries
 		$form->display();
 	}
 
   /**
    * Sets preset Values to Formfields/ -cols if a recordset is given
-   *
+   * 
+   * @todo Methode analysieren
    * @param History_DAL $p_refArticle reference to the HistoryData-Object
    * @param History $p_refHistory reference to a BcmsArticle Object holding
    * History values (optional)
    * @return int content_id
    * @author ahe
    * @date 15.11.2005 22:48:17
-   * @package
-   * @project bcms
    */
-// TODO Methode analysieren
   private function setArticlePresetValues(&$p_refArticle, $p_refHistory=null) {
 	// check whether
 	if($p_refHistory == null) {
@@ -205,15 +224,15 @@
 	  $p_refArticle->col['fk_cat']['qf_setvalue'] = $_SESSION['m_id'];
 	  return 0;
 	} else {
-		$p_refArticle->col['version']['qf_setvalue'] = (floatval($p_refHistory['version'])+1).'.0'; // URGENT die evtl. vorhandenen Versionen in der History werden hier nicht beruecksichtigt!
+		$p_refArticle->col['version']['qf_setvalue'] = (floatval($p_refHistory['version'])+1).'.0'; // \bug URGENT die evtl. vorhandenen Versionen in der History werden hier nicht beruecksichtigt!
 		$p_refArticle->col['publish_begin']['qf_setvalue'] = $p_refHistory['publish_begin'];
 	  $p_refArticle->col['publish_end']['qf_setvalue'] = $p_refHistory['publish_end'];
 	  $p_refArticle->col['fk_cat']['qf_setvalue'] = $p_refHistory['fk_cat'];
-	  $p_refArticle->col['language']['qf_setvalue'] = $p_refHistory['language'];
+	  $p_refArticle->col['lang']['qf_setvalue'] = $p_refHistory['lang'];
 	  $p_refArticle->col['heading']['qf_setvalue'] = stripslashes($p_refHistory['heading']);
 	  $p_refArticle->col['description']['qf_setvalue'] = stripslashes($p_refHistory['description']);
 	  $p_refArticle->col['layout_id']['qf_setvalue'] = $p_refHistory['layout_id'];
-	  $p_refArticle->col['status']['qf_setvalue'] = $p_refHistory['status'];
+	  $p_refArticle->col['status_id']['qf_setvalue'] = $p_refHistory['status_id'];
 	  $p_refArticle->col['prev_img_id']['qf_setvalue'] = $p_refHistory['prev_img_id'];
 	  $p_refArticle->col['prev_img_float']['qf_setvalue'] = $p_refHistory['prev_img_float'];
 	  $p_refArticle->col['redirect_url']['qf_setvalue'] = $p_refHistory['redirect_url'];
@@ -226,12 +245,14 @@
    public function createForm($contentObj,$articleId=null) {
 
 		$this->contentObj = $contentObj;
-		$histDalObj =PluginManager::getPlgInstance('ContentManager')->getModel();
+		$histDalObj = PluginManager::getPlgInstance('ContentManager')->getModel();
 		$this->updateDatabase($histDalObj);
 
-		if(($objectId=$_SESSION['current_article_data']['historyID'])>0) {
+		if(array_key_exists('current_article_data',$_SESSION)
+			&& ($objectId=$_SESSION['current_article_data']['historyID'])>0)
+		{
 			$refHistoryObj = $histDalObj->getObject($objectId,true);
-		} elseif($articleId!=null && $articleId!='') {
+		} elseif(!empty($articleId)) {
 			$artDalObj =PluginManager::getPlgInstance('ContentManager')->getArticleDalObj();
 			$refHistoryObj = $artDalObj->getObject($articleId);
 		} else {
@@ -242,23 +263,20 @@
 		$this->setArticlePresetValues($histDalObj, $refHistoryObj);
 		$contentId = ($articleId!=null && $articleId!='') ? intval($articleId) : 0;
 		// if there is an existing history object, refresh the layoutlist according to the given menu
-		if(!empty($refHistoryObj)) $histDalObj->setLayoutList($refHistoryObj['fk_cat']);
+		if(!empty($refHistoryObj)) $histDalObj->setLayoutList($refHistoryObj['fk_cat'],$this);
 
-		/* For security reasons the id of the current history record is not send to
-		 * the client as this could be manipulated.
-		 * It is read in History_DAL::checkSpecialFields()
-		 */
+		$content = (empty($refHistoryObj)) ? $refHistoryObj : unserialize($refHistoryObj['contenttext']);
 		if(!isset($_SESSION['current_article_data']))
 		{
 		  $_SESSION['current_article_data'] = array('historyID' => 0
 			,'contentID' => $contentId
-			, 'content' => unserialize($refHistoryObj['contenttext'])
+			, 'content' => $content
 			);
 		}
-		$content = unserialize($refHistoryObj['contenttext']);
-		if($refHistoryObj!=null && $content!=null)
+		if($refHistoryObj!=null && $content!=null){
 			$_SESSION['current_article_data']['content'] = $content;
-
+		}
+		
 		$_SESSION['current_article_data']['contentID'] = $contentId;
 		if(isset($_POST['layout_id'])) {
 			$curr_LayoutID = intval($_POST['layout_id']);
@@ -290,17 +308,17 @@
 
 			// show preview
 			if(isset($_POST['article_update_preview'])){
-				echo Factory::getObject('GuiUtility')->createHeading(
-					3,'Artikelvorschau',0,'previewtext_heading'); // TODO use dictionary!
+				echo BcmsFactory::getInstanceOf('GuiUtility')->createHeading(
+					3,'Artikelvorschau',0,'previewtext_heading'); // @todo use dictionary!
 				$histId = $_SESSION['current_article_data']['historyID'];
 				$this->contentObj->init('history',$histId);
 				$this->contentObj->showArticle($histId,true);
 			}
 		} elseif(isset($_POST['update_article'])) {
-			// TODO this is yet the end of the content generation process...
+			// @todo this is yet the end of the content generation process...
 			echo 'Artikel wird erstellt...';
 		} elseif(isset($_POST['article_choose_image'])) {
-			$objectmanager = PluginManager::getPlgInstance('ObjectManager');
+			$objectmanager = PluginManager::getPlgInstance('FileManager');
 			echo $objectmanager->getObjectList(true);
 		} else {
 			$this->phase1($histDalObj, 'insert_article');
